This commit was generated by cvs2svn to compensate for changes in r12325, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@12326 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/c64/reu.c
@@ -277,7 +277,7 @@ static int reu_activate(void)
         return 0;
 
     reu_ram = (BYTE *)xrealloc((void *)reu_ram, (size_t)reu_size);
-    log_message(reu_log, "%dKB unit installed.", reu_size >> 10);
+    log_message(reu_log, "%dKB unit installed.", (int)(reu_size >> 10));
 
     if (!util_check_null_string(reu_filename)) {
         if (util_file_load(reu_filename, reu_ram, (size_t)reu_size,
